Top3 feature update
Flag the top3 feature as deprecated, and advise partner to contact account manager to ensure best products are returned.
The feature wasn't not used or didn't give the expected benefit. HX account managers are happy to manually set an ideal product order in the HX side instead (at time of writing, managed in HX admin).